90% done with the specific items, need to fix the points and enchants not saving amounts/levels correctly

+++ src/main/java/com/minecave/pickaxes/menu/menus/UpgradesMenu.java
@@ -42,7 +42,7 @@ public class UpgradesMenu extends Menu {
         }
         final PItem fItem = pItem;
         int enchantCount = pItem.getEnchants().size();
-        Button[] buttons = new Button[(int) (Math.ceil(enchantCount / 2) * 9)];
+        Button[] buttons = new Button[(int) (Math.ceil((double)enchantCount / 2D) * 9)];
         fillPanes(buttons, PaneType.ALL, new FillerButton(new ItemStack(Material.AIR)));
         int c = 0;
         for (PEnchant enchant : pItem.getEnchants().values()) {
